Fix BlogView
Fix BlogView
