+++ frontend-app/src/pages/blog/edit/[edit].tsx
@@ -84,7 +84,7 @@ type BlogProps = {
   const handleSubmit = (e: FormEvent) => {
     e.preventDefault();
     // ブログ投稿後にブログ一覧ページにリダイレクト
-    router.push("/");
+    router.push("/auth/overview");
   };
 
   const handlePost = async (event: React.MouseEvent<HTMLElement>) => {
@@ -109,7 +109,7 @@ type BlogProps = {
     } finally {
       // 投稿処理が終了したので false にセット
       setEditing(false);
-      router.push("/");
+      router.push("/auth/overview");
     }
   };
 
@@ -121,7 +121,7 @@ type BlogProps = {
       <Box sx={{ p: 3 }}>
         <Box
           component={NextLink}
-          href="/"
+          href="/auth/overview"
           sx={{
             display: "inline-flex",
             height: 32,
