+++ frontend-app/src/sections/settings/settings-password.tsx
@@ -101,7 +101,7 @@ export const SettingsPassword = () => {
     } finally {
       // 処理が終了したので false にセット
       setPosting(false);
-      router.push("/");
+      router.push("/auth/overview");
     }
   };
